remove words from second string from first string
You are given two string (like two statements). You have to remove all the words of second string from first string and print the remaining first string. Please maintain the order of the remaining words from the first string. You will be only removing the first word, not all occurrence of a word. 
Example: Str1 = "A Statement is a Statement", Str2 = "Statement a" 
Output: "A is Statement"